Changed the look and feel of the HTML export file
Updated the content of the HTML (HyperText Markup Language) export file (file which contains the items/products from the shopping cart). The previous export into HTML file function was not deleted (it still exists and can be called but the new function is called instead at the pressing of the "Export" button (QPushButton element) from "Cos de cumparaturi" window).