Added JS script that creates sortable tables
The JavaScript file (script) is used to make the table containing the products/items from the shopping cart (buy list) sortable based on the values from a specific column. When the user clicks on one of the table's headers, the rows/lines/records will be sorted in ascending order on the first, third, fifth, etc. click, and in descending order on the second, fourth, sixth, etc. click. The sorting is based on the values from the column that corresponds to the clicked header.