Implemented search product in shopping cart feature
Implemented the search for product in shopping cart functionality. The method which implements the mentioned feature returns the number of products found with the specific name, type, price or producer (the user can search the product after its name, its type, its price or its producer using a combo box and a line edit) in the shopping cart's list (buy list).