Implemented data persistence for shopping cart in text file
This commit introduces the feature to save the content of the shopping cart to a text file. Each product from the buy list (shopping cart) can now be saved in the text file with its name and producer (the pair (name, producer) is unique for each product from the virtual store's catalogue). In the text file, the names and producers are separated by the '|' character.